refactor(app): move setupnetwork folder under OnDeviceDisplay
