+++ app/src/organisms/OnDeviceDisplay/SetupNetwork/FailedToConnect.tsx
@@ -13,15 +13,15 @@ import {
   Icon,
 } from '@opentrons/components'
 
-import { StyledText } from '../../atoms/text'
-import { PrimaryButton, SecondaryButton } from '../../atoms/buttons'
-import { DISCONNECT } from '../Devices/RobotSettings/ConnectNetwork/constants'
+import { StyledText } from '../../../atoms/text'
+import { PrimaryButton, SecondaryButton } from '../../../atoms/buttons'
+import { DISCONNECT } from '../../Devices/RobotSettings/ConnectNetwork/constants'
 
-import type { RequestState } from '../../redux/robot-api/types'
+import type { RequestState } from '../../../redux/robot-api/types'
 import type {
   NetworkChangeType,
   NetworkChangeState,
-} from '../Devices/RobotSettings/ConnectNetwork/types'
+} from '../../Devices/RobotSettings/ConnectNetwork/types'
 
 interface FailedToConnectProps {
   ssid: string
