+++ app/src/organisms/OnDeviceDisplay/SetupNetwork/__tests__/ConnectingNetwork.test.tsx
@@ -3,7 +3,7 @@ import { MemoryRouter } from 'react-router-dom'
 
 import { renderWithProviders } from '@opentrons/components'
 
-import { i18n } from '../../../i18n'
+import { i18n } from '../../../../i18n'
 import { ConnectingNetwork } from '../ConnectingNetwork'
 
 const render = () => {
