+++ app/src/organisms/OnDeviceDisplay/SetupNetwork/__tests__/DisplayWifiList.test.tsx
@@ -3,8 +3,8 @@ import { fireEvent } from '@testing-library/react'
 
 import { renderWithProviders } from '@opentrons/components'
 
-import { i18n } from '../../../i18n'
-import * as Fixtures from '../../../redux/networking/__fixtures__'
+import { i18n } from '../../../../i18n'
+import * as Fixtures from '../../../../redux/networking/__fixtures__'
 import { DisplayWifiList } from '../DisplayWifiList'
 
 const mockPush = jest.fn()
@@ -20,8 +20,8 @@ const mockWifiList = [
   },
 ]
 
-jest.mock('../../../redux/networking/selectors')
-jest.mock('../../../redux/discovery/selectors')
+jest.mock('../../../../redux/networking/selectors')
+jest.mock('../../../../redux/discovery/selectors')
 jest.mock('react-router-dom', () => {
   const reactRouterDom = jest.requireActual('react-router-dom')
   return {
