+++ app/src/organisms/OnDeviceDisplay/SetupNetwork/__tests__/SelectAuthenticationType.test.tsx
@@ -4,8 +4,8 @@ import { fireEvent } from '@testing-library/react'
 
 import { renderWithProviders } from '@opentrons/components'
 
-import { i18n } from '../../../i18n'
-import * as Networking from '../../../redux/networking'
+import { i18n } from '../../../../i18n'
+import * as Networking from '../../../../redux/networking'
 import { SetWifiCred } from '../SetWifiCred'
 import { SelectAuthenticationType } from '../SelectAuthenticationType'
 
@@ -14,9 +14,9 @@ const mockSetShowSelectAuthenticationType = jest.fn()
 const mockSetSelectedAuthType = jest.fn()
 const mockSetChangeState = jest.fn()
 
-jest.mock('../../../organisms/SetupNetwork/SetWifiCred')
-jest.mock('../../../redux/networking')
-jest.mock('../../../redux/discovery/selectors')
+jest.mock('../SetWifiCred')
+jest.mock('../../../../redux/networking')
+jest.mock('../../../../redux/discovery/selectors')
 jest.mock('react-router-dom', () => {
   const reactRouterDom = jest.requireActual('react-router-dom')
   return {
