+++ app/src/organisms/OnDeviceDisplay/SetupNetwork/__tests__/SetWifiSsid.test.tsx
@@ -4,8 +4,8 @@ import { fireEvent } from '@testing-library/react'
 
 import { renderWithProviders } from '@opentrons/components'
 
-import { i18n } from '../../../i18n'
-import { JOIN_OTHER } from '../../Devices/RobotSettings/ConnectNetwork/constants'
+import { i18n } from '../../../../i18n'
+import { JOIN_OTHER } from '../../../Devices/RobotSettings/ConnectNetwork/constants'
 import { SetWifiSsid } from '../SetWifiSsid'
 
 const mockSetSelectedSsid = jest.fn()
